+++ tests/testthat/test-AdditionalCriteria.R
@@ -28,14 +28,14 @@ test_that("cohort with additional criteria works", {
   cohortList <- toCirce(cd)
   expect_type(cohortList, "list")
 
-  cohortJson <- jsonlite::toJSON(cohortList, pretty = T, auto_unbox = TRUE) %>%
+  cohortJson <- jsonlite::toJSON(cohortList, pretty = T, auto_unbox = TRUE) |>
     as.character()
 
   expect_type(cohortJson, "character")
   expect_true(nchar(cohortJson) > 1)
 
 
-  sql <- CirceR::cohortExpressionFromJson(cohortJson) %>%
+  sql <- CirceR::cohortExpressionFromJson(cohortJson) |>
     CirceR::buildCohortQuery(options = CirceR::createGenerateOptions(generateStats = TRUE))
 
   expect_type(sql, "character")